Avatar/requirements.txt
# Core
numpy==1.23.4
scipy==1.10.1
pillow==10.2.0
pyyaml
joblib==1.1.0
tqdm
requests==2.31.0

# Image & Video Processing
opencv-python==4.10.0.82
scikit-image==0.19.3
imageio==2.19.3
imageio-ffmpeg==0.4.7
ffmpeg-python==0.2.0
pydub==0.25.1
av

# Audio Processing
librosa==0.9.2
numba
resampy==0.3.1

# Face & Image Enhancement
face-recognition==1.3.0
face-alignment==1.3.5
mediapipe==0.10.14
kornia==0.6.8
basicsr==1.4.2
facexlib==0.3.0
gfpgan

# Text-to-Speech
edge-tts==6.1.10

# Background Removal
rembg[gpu]==2.0.51

# Gradio UI
gradio

# Misc
yacs==0.1.8
safetensors
